Implement .read_digits for multiple OCR strings
Building on our .scan_number method, we use that in .read_digits,
a method that iterates over the string of OCR characters, parsing
each 3x3 block from pipe/underscore to an actual number represented
as a string.

In the interest of time, I did not spec out all of the OCRs in the
sample.txt file.