Implement .parse_file method
Taking a txt file of OCR strings, this method reads the file and
passes the appropriate blocks of pipe/underscore characters to
the methods we previously implemented. I have decided to store the
list in an array, which could be output in any number of ways.

Part of future cleanup/refactor might be to create a new file
with the parsed numbers. I would also like to write RDoc for
each method to add some detail to each method.