Add various comments, error handling
This last branch simply attempts to leave the project in a decent
state, one with some explanation of the methods and a comment or two
on things I'd like to improve (or would keep in mind for future
improvemnet based on changing requirements).

As it's getting late, I hope that we can continue the discussion
over a pairing session and I'm really looking forward to having
the chance to spend some time reviewing this with you. Cheers!